+++ src/dev/Renderer/index.js
@@ -11,7 +11,7 @@ import { Renderer as PrismaCmsRenderer } from '@prisma-cms/front'
 import Context from "@prisma-cms/context";
 
 import MainMenu from './MainMenu';
-import { withStyles } from 'material-ui';
+import withStyles from 'material-ui/styles/withStyles';
 import DevMainPage from './pages/MainPage';
 import TemplatesPage from '../../components/pages/Templates';
 import TemplatePage from '../../components/pages/Templates/Template';